requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
pandas==2.1.3
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
fastapi
uvicorn
jinja2
pandas-ta
python-multipart
httpx